rename task name to lambda and more test for lambda
